Fixed SaveNoteInteractorTest.java and refactored SaveNote use cases tests
